// packages/admin/src/generated/graphql.tsx
export type Maybe<T> = T | null
// ...
export type Scalars = {
  ID: { input: string; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  Upload: { input: any; output: any }
}

export type AdviceType = {
  __typename?: 'AdviceType'
  category?: Maybe<Scalars['String']['output']>
  createdAt?: Maybe<Scalars['String']['output']>
  defaultValue?: Maybe<Scalars['String']['output']>
  id: Scalars['Int']['output']
  indexNum?: Maybe<Scalars['Int']['output']>
  lastModifiedTime?: Maybe<Scalars['String']['output']>
  onOff?: Maybe<Scalars['String']['output']>
  studentStates?: Maybe<Array<Maybe<StudentState>>>
  type: Scalars['String']['output']
  updatedAt?: Maybe<Scalars['String']['output']>
}
// ...
export type Branch = {
  __typename?: 'Branch'
  branchName?: Maybe<Scalars['String']['output']>
  createdAt: Scalars['String']['output']
  id: Scalars['Int']['output']
  lastModifiedTime?: Maybe<Scalars['String']['output']>
  updatedAt: Scalars['String']['output']
}
// ...
export type ConsultationMemo = {
  __typename?: 'ConsultationMemo'
  Branch?: Maybe<Branch>
  branchId?: Maybe<Scalars['Int']['output']>
  content: Scalars['String']['output']
  createdAt: Scalars['String']['output']
  id: Scalars['Int']['output']
  lastModifiedTime?: Maybe<Scalars['String']['output']>
  manageUser?: Maybe<ManageUser>
  manageUserId?: Maybe<Scalars['Int']['output']>
  studentState?: Maybe<StudentState>
  studentStateId?: Maybe<Scalars['Int']['output']>
  updatedAt: Scalars['String']['output']
}
// ...
export type ManageUser = {
  __typename?: 'ManageUser'
  Branch?: Maybe<Array<Maybe<Branch>>>
  ConsultationMemo?: Maybe<Array<Maybe<ConsultationMemo>>>
  Lectures?: Maybe<Array<Maybe<Lectures>>>
  PaymentDetail?: Maybe<Array<Maybe<PaymentDetail>>>
  Stamp?: Maybe<Array<Maybe<Stamp>>>
  Student?: Maybe<Array<Maybe<Student>>>
  StudentMemo?: Maybe<Array<Maybe<StudentMemo>>>
  StudentPayment?: Maybe<Array<Maybe<StudentPayment>>>
  StudentStates?: Maybe<Array<Maybe<StudentState>>>
  branchId?: Maybe<Scalars['Int']['output']>
  createdAt?: Maybe<Scalars['String']['output']>
  email?: Maybe<Scalars['String']['output']>
  favoriteStudentState?: Maybe<Array<Maybe<Scalars['Int']['output']>>>
  frequentlyUsed?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  id: Scalars['Int']['output']
  lastModifiedBy?: Maybe<Scalars['String']['output']>
  lastModifiedTime?: Maybe<Scalars['String']['output']>
  mAddressDetail?: Maybe<Scalars['String']['output']>
  mAddresses?: Maybe<Scalars['String']['output']>
  mAvatar?: Maybe<Scalars['String']['output']>
  mGrade?: Maybe<Scalars['Int']['output']>
  mJoiningDate?: Maybe<Scalars['String']['output']>
  mPart?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  mPassword: Scalars['String']['output']
  mPhoneNum?: Maybe<Scalars['String']['output']>
  mPhoneNumCompany?: Maybe<Scalars['String']['output']>
  mPhoneNumFriend?: Maybe<Scalars['String']['output']>
  mPhoneNumInside?: Maybe<Scalars['String']['output']>
  mRank?: Maybe<Scalars['String']['output']>
  mUserId: Scalars['String']['output']
  mUsername: Scalars['String']['output']
  mZipCode?: Maybe<Scalars['String']['output']>
  resign?: Maybe<Scalars['String']['output']>
  updatedAt?: Maybe<Scalars['String']['output']>
}
// ...
export type StudentState = {
  __typename?: 'StudentState'
  Branch?: Maybe<Branch>
  adviceTypes?: Maybe<Array<Maybe<AdviceType>>>
  agreement: Scalars['String']['output']
  branchId?: Maybe<Scalars['Int']['output']>
  campus?: Maybe<Scalars['String']['output']>
  category?: Maybe<Scalars['String']['output']>
  classMethod?: Maybe<Array<Maybe<Scalars['String']['output']>>>
  consultationMemo?: Maybe<Array<Maybe<ConsultationMemo>>>
  createdAt: Scalars['String']['output']
  currentManager?: Maybe<Scalars['String']['output']>
  detail?: Maybe<Scalars['String']['output']>
  expEnrollDate?: Maybe<Scalars['String']['output']>
  id: Scalars['Int']['output']
  lastModifiedTime?: Maybe<Scalars['String']['output']>
  perchase?: Maybe<Scalars['Boolean']['output']>
  phoneNum1: Scalars['String']['output']
  phoneNum2?: Maybe<Scalars['String']['output']>
  phoneNum3?: Maybe<Scalars['String']['output']>
  pic?: Maybe<Scalars['String']['output']>
  progress: Scalars['Int']['output']
  receiptDiv?: Maybe<Scalars['String']['output']>
  stAddr?: Maybe<Scalars['String']['output']>
  stEmail?: Maybe<Scalars['String']['output']>
  stName: Scalars['String']['output']
  stVisit?: Maybe<Scalars['String']['output']>
  subDiv?: Maybe<Scalars['String']['output']>
  subject: Array<Maybe<Scalars['String']['output']>>
  updatedAt: Scalars['String']['output']
}